before removing Besim code

+++ src/org/sunspotworld/Rx_package.java
@@ -13,10 +13,10 @@ public class Rx_package {
   private int pck_type = -1;
   private int node_index = -1;
   private String dst_addr = null;
-  private int[] payload = null;
+  private short[] payload = null;
 
   public Rx_package(int pck_type, int node_index,
-                    String dest_addr, int[] data)
+                    String dest_addr, short[] data)
   {
     this.pck_type = pck_type;
     this.node_index = node_index;
@@ -24,7 +24,7 @@ public class Rx_package {
     if(dest_addr != null)
       this.dst_addr = new String(dest_addr);
     
-    this.payload = new int[data.length];
+    this.payload = new short[data.length];
     System.arraycopy(data, 0, this.payload, 0, data.length);    
   }
 
@@ -40,11 +40,11 @@ public class Rx_package {
     return this.dst_addr;
   }
 
-  public int[] get_payload() {
+  public short[] get_payload() {
     return this.payload;
   }
 
-  public void set_payload(int i, int val) {
+  public void set_payload(int i, short val) {
     payload[i] = val;
   }
 }
